fix: 0612 역 기준 식당 검색 피드백 수정

+++ src/main/java/com/honbabmap/backend/restaurant/entity/RestFeatEntity.java
@@ -12,8 +12,7 @@ public class RestFeatEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "rest_feat_id")
-    private Long restFeatId;
-
+    private Integer restFeatId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "rest_id")
